requirements.txt
flask
tempfile
numpy
pillow
opencv-python